requirements.txt
pyserial==3.5
matplotlib==3.3.2
numpy==1.21.3
pyinstaller-hooks-contrib>=2021.0
